看板描画

+++ src/user/Others/HandShake.h
@@ -12,14 +12,16 @@ class HandShake
 	Vec2<int>m_seed = { 0,1 };
 
 public:
-	int m_interval = 300;
+	float m_interval = 300.0f;
 	Vec2<float>m_val = { 1.0f,1.0f };
 
 	HandShake() {}
 	void Init();
 	void Update(float arg_timeScale,const Matrix& arg_rotate);
 
+	//ƒuƒŒ—Ê
 	const Vec3<float>& GetOffset()const { return m_offset; }
+	//¬”“_•”•ª
 	float GetTimeRate()const { return m_t - static_cast<float>(static_cast<int>(m_t)); }
 };
